style: 코드 컨벤션에 맞게 컴포넌트 파일 분리
